[Refactor] Opt common module
