+++ be/src/exec/pipeline/pipeline_driver_poller.cpp
@@ -16,10 +16,10 @@
 
 #include <chrono>
 
-#include "common/logconfig.h"
 #include "exec/pipeline/pipeline_fwd.h"
 #include "exec/pipeline/pipeline_metrics.h"
 #include "runtime/exec_env.h"
+#include "runtime/logconfig.h"
 #include "util/time_guard.h"
 
 namespace starrocks::pipeline {
