+++ be/src/exec/pipeline/schedule/timeout_tasks.cpp
@@ -14,10 +14,10 @@
 
 #include "exec/pipeline/schedule/timeout_tasks.h"
 
-#include "common/logconfig.h"
 #include "exec/pipeline/fragment_context.h"
 #include "exec/pipeline/pipeline_fwd.h"
 #include "exec/pipeline/schedule/common.h"
+#include "runtime/logconfig.h"
 #include "util/stack_util.h"
 
 namespace starrocks::pipeline {
