+++ be/src/exec/schema_scanner/schema_be_logs_scanner.h
@@ -16,9 +16,9 @@
 
 #include <cstdint>
 
-#include "common/greplog.h"
 #include "exec/schema_scanner.h"
 #include "gen_cpp/FrontendService_types.h"
+#include "service/greplog.h"
 
 namespace starrocks {
