+++ be/src/http/action/pprof_actions.cpp
@@ -41,7 +41,6 @@
 #include <mutex>
 
 #include "common/config.h"
-#include "common/prof/heap_prof.h"
 #include "common/status.h"
 #include "common/tracer.h"
 #include "http/ev_http_server.h"
@@ -49,6 +48,7 @@
 #include "http/http_headers.h"
 #include "http/http_request.h"
 #include "io/io_profiler.h"
+#include "runtime/prof/heap_prof.h"
 
 namespace starrocks {
