+++ be/src/runtime/closure_guard.h
