+++ be/src/runtime/load_channel_mgr.cpp
@@ -40,9 +40,9 @@
 #include <memory>
 
 #include "base/concurrency/stopwatch.hpp"
-#include "common/closure_guard.h"
 #include "fs/key_cache.h"
 #include "gutil/strings/substitute.h"
+#include "runtime/closure_guard.h"
 #include "runtime/exec_env.h"
 #include "runtime/load_channel.h"
 #include "runtime/mem_tracker.h"
