+++ be/src/runtime/local_tablets_channel.cpp
@@ -26,12 +26,12 @@
 #include "base/concurrency/stopwatch.hpp"
 #include "base/string/faststring.h"
 #include "column/chunk.h"
-#include "common/closure_guard.h"
 #include "common/statusor.h"
 #include "exec/tablet_info.h"
 #include "gen_cpp/internal_service.pb.h"
 #include "gutil/ref_counted.h"
 #include "gutil/strings/join.h"
+#include "runtime/closure_guard.h"
 #include "runtime/descriptors.h"
 #include "runtime/exec_env.h"
 #include "runtime/global_dict/types.h"
