+++ be/src/service/CMakeLists.txt
@@ -22,6 +22,8 @@ add_subdirectory(service_be)
 set(EXEC_FILES
     backend_options.cpp
     backend_base.cpp
+    daemon.cpp
+    greplog.cpp
     internal_service.cpp
     staros_worker.cpp
     mem_hook.cpp
