+++ be/src/service/internal_service.cpp
@@ -50,7 +50,6 @@
 #include "brpc/errno.pb.h"
 #include "cache/datacache.h"
 #include "column/stream_chunk.h"
-#include "common/closure_guard.h"
 #include "common/compiler_util.h"
 #include "common/config.h"
 #include "common/process_exit.h"
@@ -71,6 +70,7 @@
 #include "gutil/strings/substitute.h"
 #include "runtime/batch_write/batch_write_mgr.h"
 #include "runtime/buffer_control_block.h"
+#include "runtime/closure_guard.h"
 #include "runtime/command_executor.h"
 #include "runtime/data_stream_mgr.h"
 #include "runtime/descriptors.h"
