+++ be/src/service/service_be/internal_service.cpp
@@ -34,13 +34,13 @@
 
 #include "internal_service.h"
 
-#include "common/closure_guard.h"
 #include "common/config.h"
 #include "common/utils.h"
 #include "exec/pipeline/fragment_context.h"
 #include "gen_cpp/BackendService.h"
 #include "gutil/strings/substitute.h"
 #include "runtime/buffer_control_block.h"
+#include "runtime/closure_guard.h"
 #include "runtime/data_stream_mgr.h"
 #include "runtime/fragment_mgr.h"
 #include "runtime/load_channel_mgr.h"
