+++ be/src/service/service_be/starrocks_be.cpp
@@ -24,7 +24,6 @@
 #include "cache/datacache.h"
 #include "cache/disk_cache/block_cache.h"
 #include "common/config.h"
-#include "common/daemon.h"
 #include "common/process_exit.h"
 #include "common/status.h"
 #include "fs/s3/poco_common.h"
@@ -35,6 +34,7 @@
 #include "runtime/jdbc_driver_manager.h"
 #include "service/backend_options.h"
 #include "service/brpc.h"
+#include "service/daemon.h"
 #include "service/service.h"
 #include "service/service_be/arrow_flight_sql_service.h"
 #include "service/service_be/http_service.h"
