+++ be/src/service/starrocks_main.cpp
@@ -55,7 +55,6 @@
 #include "agent/heartbeat_server.h"
 #include "agent/status.h"
 #include "common/config.h"
-#include "common/daemon.h"
 #include "common/logging.h"
 #include "common/process_exit.h"
 #include "common/status.h"
@@ -65,6 +64,7 @@
 #include "runtime/jdbc_driver_manager.h"
 #include "runtime/memory/roaring_hook.h"
 #include "service/backend_options.h"
+#include "service/daemon.h"
 #include "service/service.h"
 #include "service/staros_worker.h"
 #include "storage/options.h"
