+++ be/src/util/cow.h
