+++ be/test/util/cow_test.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "common/cow.h"
+#include "util/cow.h"
 
 #include <glog/logging.h>
 #include <gtest/gtest.h>
